Install selected theme during settings persistence.
Themes have to be installed to be usable by the system. This is not done
automatically which lead to the problems with the site when themes where
switched right after the install and before the "Reload themes" button
was pressed in the admin page.

This patch makes sure that the selected them is always installed when
it is selected by the user either in the admin or settings page.

Fixes #4921.