Vagrant: Do not pull PHP requirements during VM setup
Don't pull the PHP requirements of Friendica with composer during the VM setup.

First coding is done out side of the VM, so should be the management of the
dependencies. Additionally if the _VirtualBox Guest Additions_ versions don't
match on the guest and host systems, writing to the shared directory will not
work, thus running composer will generate error messages during the initial
setup and Friendica wont work.