show "parent unknown" when parent not federated
in some cases (visibility is private of parent comment), there is no
parent federated to my instance, and then the comment "hangs around"
with no information, why there is no parent.

For normal users this can be confusing, why all exept "this" comment
have parents.

So i added a simple text instead of link to parent with a "title" field
with a small explanation, why parent is not viewable.

translations done